Fix CI errors
